Updated openstack/openstack Project: openstack/neutron  c83fb1064c58ab74572d948268d61133dc485e12
Fix tenant-id in Arista ML2 driver to support HA router

When HA router is created, the framework creates a network and does
not specify the tenant-id. This casuse Arista ML2 driver to fail.
This patch sets the tenant-id when it is not passed explicitly by
by the network_create() call from the HA router framework.

Even though original issue was discovered for network_create()
it turns out the same issue applies for port_create() as well
so all the methods are fixed in this patch

Change-Id: Id03e26dac003fa589477ed78a4ed3c44cc2f708d
Closes-Bug: 1468828