Updated openstack/openstack Project: openstack/neutron  cde607e05d2f05495699d7d38c6723085c667ef7
Don't process network_delete events on OVS agent

There is no reason to reclaim the local VLAN on a network_delete
event since it will be reclaimed when the last port is deleted anyway.

This method was racey with recent port creation events and lead to
occasional traces when it would yank the local VLAN out from underneath
an ongoing port wiring process.

Change-Id: I5eee6175c053006b35b6efe274d27931a9d5d89f
Closes-Bug: #1468251