Updated openstack/openstack Project: openstack/tempest  63e51066b09da86add9c6b94f02a000b3848fd61
Re-factor neutron client for 'create' methods

Currently neutron client is written to auto generate
methods on fly. This is very difficult to debug and
make code very difficult to read.

As neutron client will be moved to lib, it will be better to
have its interfaces more readable and consistent with other
service clients.

This commit re-factor it for 'create_*' methods.

Change-Id: I77cec3cdb1bdf404fab6083c1417368bc8fb6cf5