Updated openstack/openstack Project: openstack/oslo.db  eeacae192e69f56b38e5cfa0654e3a653bc9258a
Allow additional exceptions in wrap_db_retry

There are cases where we want to catch and retry some additional
exceptions and treat them as RetryRequests depending on method
that is being excecuted (e.g. StaleDataError or ObjectDeletedError).

This patch changes the wrap_db_retry decorator to take an extra
parameter that is a list of additional exception classes that will
be caught and treated as a RetryRequest.

Change-Id: I6c61b1d2f0331cac31dfde0afa523e44d470b83f