Updated openstack/openstack Project: openstack-infra/releasestatus  4d5c95e8145399ea6034a3f206480430e89dbb34
Loop to retrieve all reviews

releasestatus was only making a single query to list all reviews
across all projects, so in those busy days it was missing a whole
lot of review links.

This change introduces per-project queries and sortKey-based loops
to retrieve all the related changes.

Also drops retrieving is:submitted since that is not used at all.

Change-Id: I3d4abc8bc70904ae7c599d8efa3fac956aab702b