Updated openstack/openstack Project: openstack/tuskar-ui  a57fa800542ed2ee61b931bc6d2991eca0d946e6
Properly filter Ceilometer statistics for charts

We used to filter by image_id for roles, back when we had different
images for different roles. As we are now using the same image for
all the roles, that way of filtering does not work any longer. We
need to filter by instance UUIDs instead.

Change-Id: Ie4f26a29beaaf67aa3b36bb82961b491d31c95a0