Updated openstack/openstack Project: openstack-dev/devstack  e4d57d6bed88ec11ca7fcc8dc286a69d284c70aa
Fixes for Linux Bridge in the L3 agent

Prior to this patch, the logic for configuring the interface used for
the L3 agent was OVS specific. This patch introduces code to correctly
identify the brq device that is used for the L3 agent when using the
Linux Bridge mechanism driver.

Change-Id: I1a36cad0fb790aaa37417a1176576293e4f2c87f
Co-Authored-By: Jens Rosenboom <j.rosenboom@x-ion.de>