Updated openstack/openstack Project: openstack/swift  6247508ae941e68982b446bd6ba39b2a2aeb738b
Fix ValueError in ssync_receiver

httplib's putheader method will cast whatever you give it to a string.
where we allow the default dict.get default of None to be passed to
putheader unmodified ssync_receiver is surpised that the non-empty
string isn't able to be converted to an integer.

We can avoid surprising the ssync_receiver in this way by sending the
empty string as a better default.

Change-Id: Ie9df9927ff4d3dd3f334647f883b2937d0d81030