Updated openstack/openstack Project: openstack/trove  3fca3fcb96559178c2ce94ad5ca4cf26e4e4a7be
Root enablement for Vertica clusters/instances

Adds support for root-enable for Vertica clusters,
with optional user-provided passwords. In order to
support root-enable for clusters, a new route was added.

A new VerticaRootController extends the features that
the RootController provides for clusters. This datastore
specific controller is added to the subfolder vertica
in extensions.

Added unit tests for new Vertica functionalities.

Refactoring:
1. Made this patch dependant on pxc patch, in order to be able to
address these changes for pxc clustering. Also added
rootController entry to pxc's config

2. Moved all the RootEnable code inside trove/extensions/common.
Since there was nothing specific to mysql in that code.

3. Changed the cfg.py to now include a default RootController for
all the datastores. Because without this patch all the datastores
were going through the same old mysql root controller anyways,
so no need to change the bahavior of the other datastores. Which
ever datastores don't provide root implementation they will raise
appropriate exceptions anyways.

4. If root_password is not passed then calling enable_root, if the
password is supplied then calling enable_root_with_password. That way
the datastores which provide enable_root implementation but not
enable_root_with_password can be handled correctly, appropriate
exceptions are raised to indicate to user that supplied password
will not be honored.

5. If root-enable is called on clusters, then default root controller
implementation raises exception, as other than vertica nobody
provides the implementation for it as of now. For vertica, this
is processed correctly.

Change-Id: I28209ab562b26f45742b5db65c5cc057e07f7f51
Implements blueprint: vertica-cluster-user-features
Depends-On: I3e23c87c0e48a7f910e54207116fc75b5ce0ad0c