Updated openstack/openstack Project: openstack/sahara  f0325c46435f3a8caf8528de132de2b58c01c9dd
Modify launch_command to support global variables

We use launch_command script to run Spark and Storm jobs on the guest.
Sometimes like in case of the Spark Shell job for the proper job execution
we need to set environment parameters on the guest before running the job.
launch_command is using the subprocess module to run the command.
subprocess.Popen is already supporting environment variables through the
"env" parameter, so all we have to do is populate this variable if needed.

This patch proposes to use the same format to transfer environment variables
to the command we use when running a shell script:

FILE="a" DIR="b" test.sh

In this case launch_command will parse a command line, separate environment
variables FILE and DIR from the command and send them to Popen using the "env"
parameter.

Change-Id: If110f00e95f0479b12e2d26cc30e9cbe4513e361