Updated openstack/openstack Project: openstack/ceilometer  d0a588aa95af57add95c368f119e67d085424ce6
restore long uuid data type

partial revert of Iffd6001c337fa692cc3925778eeb98dda46b2c78.
we are not able to revert the patch to support those chasing trunk.
this patch makes original migration a noop for future cases and
adds a follow up script to revert migration if already applied.

it does not revert ids that are under the domain of ceilometer and
are known to be within 128 char length, specifically:

- message_id
- message_signature
- event_id (alarm)
- alarm_id

Change-Id: I8c191adb897f27a1ce0f199ed35253324a5c3809
Closes-Bug: 1468916