Updated openstack/openstack Project: openstack/heat  2ed0b7ddcbb47b28d3a0cc8b2831fc2e51b59e2d
Convergence rollback implementation

When an error occurs while provisioning a resource and rollback for the
stack is enabled, then rollback is triggered.

Stack table has reference to previous completed template and it is used
for issuing update on stack in order to rollback. If previous template
doesn't exist then an empty template is supplied (this is the case of
create rollback).

Change-Id: I3c9ef8665cc3b2006b8d0c330b00fa5e226201b9
Implements: convergence-rollback