Updated openstack/openstack Project: openstack/cinder  e6f61b9d5bcdfbf2d833879c38b7d99e11ab5444
Use symbol for error code in VNX cinder driver

Some of the error code of VNX CLI is scattered in the code and is
not coded as a constant.

Gather all VNX error codes in the VNXError class and compose a new
method to check the error in the CLI output.  If no specific error
is supplied, check all available errors in the output.

Change-Id: I145803fbe9fc36dcf432c9aa4d90245790a69859
Closes-Bug: #1454905