Updated openstack/openstack Project: openstack/requirements  bddbdc35328233aa7b72ec9364ff943489db1534
Prevent incompatible version bumps.

We shouldn't be able to bump global requirement versions if that
would be incompatible with the current test set of versions. To assess
'works' we need to run them in tests which means that we need to
update upper-constraints.txt at the same time. We don't test all
versions however, so just checking that the constraints version is
compatible with the updated global requirements is sufficient.

This adds a dependency on the pypa packaging library. The library is
maintained by the authors of setuptools and pip, is Apache V2, in good
condition and Python 3 ready. Its a test-only dependency so I don't
think the distro packaging status matters. This is *the* reference
library for packaging facilities on Python.

Change-Id: I032dd9a384453dcfa4911ea905c8f89b0c974211