Updated openstack/openstack Project: openstack/cinder  09a213facccaed6358581cf92d3c732f02b9feac
Remove generate_glance_url

The function generate_glance_url is part of the inital fork from nova.
It is only called once and that is by a unit test designed to test the
function.

This commit removes generate_glance_url and its associated test because
they are not used by cinder.

Change-Id: I02a3632d258b7555284e66368eaf17af3b4b0a2f