Updated openstack/openstack Project: openstack/heat  fc529025e9f38e0b58908db74e8d7f3999195186
Allow multiple lists for list_join intrinsic functions

Allow the HOT list_join function to optionally accept multiple
lists.  This is useful when you want to combine two or more lists,
for example to join a hard-coded list with a user provided one.

Although this is 100% backwards compatible, it's been added as a new
function which is available only via the Liberty HOT version.  This
should allow easier determination of whether a particular heat service
supports this revised syntax.

Change-Id: I375450960c6d486998308ec0b915c25628f745a5