Updated openstack/openstack Project: openstack/neutron  8942fccf02e6e179d47582fdb2792a1ca972da21
Remove failing SafeFixture tests

The fixtures 1.3 release attempted to fix the fixtures resource
leak issue, but failed to do so completely. Our own SafeFixture
is still needed: The 1.3 release broke our SafeFixture tests,
but not the usage of SafeFixture itself. This patch removes
those failing tests for now to unbreak the gate. Jakub reported
a bug on fixtures 1.3:
https://bugs.launchpad.net/python-fixtures/+bug/1469759

We will continue to use SafeFixture until that bug is fixed
in fixtures, at which point we will be able to require
fixtures > 1.3.

Change-Id: I59457c3bb198ff86d5ad55a1e623d008f0034b8f
Closes-Bug: #1469734