Updated openstack/openstack Project: openstack-infra/infra-manual  873d2dba0d289407a113f3b38f309eff6348f722
Add links to recorded sessions on debugging the gate to developer docs

While trying to explain to someone the process of debugging automated
test failures in the gate they said they had a hard time finding good
docs that explain the process.

The developer docs have some high level steps but since we have recorded
sessions on this topic we should link them in since going through real
examples and showing people how to do it in a recording should also be
helpful for education.

Change-Id: Ieea6915b7224163fa005538c44280389eed51659