Updated openstack/openstack Project: openstack-infra/project-config  8e764f9ed47bc1f73258381e08dd8ece08d8bf1b
Use the requirements parser.

This switches the requirements parser from a local one to using the
reference one in openstack/requirements. To do that I install the
openstack/requirements we're testing against in a virtualenv and
import it. We could if wanted install that separately and run it
against a just in time checkout, but this should be reliable for now
(though the need to encode a python path in a venv is a little ugly).

Change-Id: I1f77a2e0eaa7046928642cf0b9e2c490a851a303