Updated openstack/openstack Project: openstack/designate  77783cb3776c387039980090f4768bc83e4c3cee
Allow logging hookpoint intialization

The hook points are applied prior to any logging configuration. This
makes it difficult for a hook author to ensure that the correct
configuration and package entry points have been added. This patch takes
the log messages and buffers them for replay after the logging has been
configure. The replay function is then called in each of the commands
after the configuration has been read.

Closes-Bug: #1465731
Change-Id: Ia5cc2dba53056c55017b3a104a1f4e22e6594ae0