Updated openstack/openstack Project: openstack/neutron  ce65573d4e424462d60acfd73f1531e0705720ed
Retry port status update on StaleDataError

During update_port_status, a port may be concurrently updated/deleted
by another process, which will result in a StaleDataError being
thrown. In the event it was an update, we want to retry to make sure
the port status is set to the desired state so third parties like Nova
get notified.

This patch uses the oslo_db wrap_db_retry decorator after converting
the StaleDataError into a RetryRequest that the decorator can catch.

Closes-Bug: #1465407
Change-Id: I0a9230d30f435ec3d48f033136e85c40ad489ef9