Updated openstack/openstack Project: openstack/tempest  8ce7024a2c6f405175c9bd77043f8b2fd87458fc
Add compute_feature_enabled.attach_encrypted_volume config option

TestEncryptedCinderVolumes passes for the ceph job today but it's a
false positive since the rbd volume driver in cinder does not return the
'encrypted' key in it's connection_info dict from the
os-initialize_connection API to Nova, and Nova keys off 'encrypted' in
connection_info to see if it should run the volume encryption provider
when attaching the volume (in the case of the libvirt driver in Nova).

Cinder change I03f8cae05cc117e14f7482115de685fc9f3fa54a sets the
'encrypted' key in connection_info for rbd volumes which then makes Nova
attempt volume encryption but that fails for the rbd volume type since
it's not currently supported in Nova. Eventually the tests fail in
Tempest because the volume status does not go to 'in-use' since the
attach failed.

This change adds a config option so that the encrypted cinder volume
tests can be skipped in the ceph job until rbd volume encryption is
supported in Nova.

An alternative to a new config option would be to check if the
CONF.volume.storage_protocol is 'ceph' and raise a skip exception for
bug 1463525, but given the number of other cinder volume drivers that
might have this same issue I figured it was best to make Tempest
configurable rather than hard-code all of the invalid storage protocols
in the test.

Related-Bug: #1463525

Change-Id: I48eba7c645cc1c979fd766ae9c05efb00957f787