Updated openstack/openstack Project: openstack-infra/project-config  1d333833bb0958edd4602d1997ca3a814a4626f1
Add beaker-localhost and beaker-nodepool hypervisors

Beaker(puppet testing tool) has plugable hypervisor backends. Create one
project to manage the 'ssh to localhost' backend and one to hit the
nodepool api (rest or gear) remotely.

The aviator builders are here because they are actually just good solid
ruby tests. If this works out we can look at refactoring some generic
ruby test builders.

The beaker-localhost hypervisor backend can be used instead of the None
we are using now. This is more hygenic because we can specify exactly
how to behave instead of counting on fallback behaviour that may change.
There are some things this wont fix such as using ssh at all and
managing sshd_config. I've been talking about writing our own hypervisor
for several weeks now so I'd like to at least get it stubbed out as a
potential extention point.

The beaker-nodepool hypervisor might better live in stackforge/. It is a
hypervisor for beaker that interacts with the nodepool service to
provision nodes. I will likely use this with a standalone nodepool (no
zuul) for another project. It makes sense to keep this close to the
nodepool community(infra) but that isn't a technical requirement.

Change-Id: Iaad72a26b26a0ee89c4fd254d0bcb75535a8d551