Updated openstack/openstack Project: openstack/keystonemiddleware  661c19cde75fd8e40ab24f65eebadd3ba72a8795
Refactor certificate fetch functions

The certificate fetch function code was complicated because you
passed in a string to pick the certificate type. This was a
holdover from before keystoneclient was used because it was
building the paths to the certificates. Now auth_token is just
calling client methods rather then building the path so the strings
don't mean anything anymore.

Change-Id: Iede0c00039dacb7ff3460d7af38151d0f6f8c433