Updated openstack/openstack Project: openstack-infra/project-config  e288a618a41f765525d67ba7fd8fe4dfab021232
run-docs.sh: execute docs job if present

Some projects (like neutron) provide their own docs job, so we should
use it instead of running build_sphinx directly.

The use case to maintain a custom docs job is f.e. to enable sphinx
warnings in gate. They were enabled before for neutron under assumption
that gate relies on the job, but it's not the case, as it turned out.

Note that projects that want to expose the tox job to gate should have
it defined in tox envlist so that 'tox -l' returns it.

Depends-On: I80eb169b7b4e5a3490586722c64394dbb724928d
Depends-On: Id276fa59edb33f7789ab06055300b4dc2385472a
Depends-On: I8c265eae2175425568479116d1faef7d87fdcc02
Change-Id: I433126a8247e7e1c316f2c96bb21e15582b247ce