Updated openstack/openstack Project: openstack-infra/project-config  2c9678d5279a781d9c79c8fb2b2fab87ee981159
Ensure we have py2.6 jobs enabled for cloud-init

The goal of cloud-init is to support as many versions
of systems as possible and that includes rhel6 or centos6
which are still actively in use. Those systems run py2.6
so we should continue to run on these systems to ensure we
are still compatible.

Change-Id: I66f5fff3c821b143df00fc51f03bd714895aec57