Updated openstack/openstack Project: openstack-infra/system-config  7577526b20da65b685c2174332b750e2929b203b
Trim user specific package a bit

I'm pretty sure I'm the reason byobu is on any of our servers, and I'm
also pretty sure I haven't used it in at least 2 years.

Also - starting with 15.04, Ubuntu has a vanilla emacs-nox like redhat
does. Let's start using it so that we don't have to do patches like this
in the future.

Change-Id: I37123dd42bdbd6456bb8a8637ab7164af7132642