Updated openstack/openstack Project: openstack/python-cinderclient  167a75156de71aa2c1dcf963d76b9a08cee33bbf
Set max volume size limit for the tenant

There is a need to limit maximum size of a volume to levels
that the storage infrastructure can handle.
Setting a maximum limit on size of a volume also prevents
a tenant from creating large volumes that have not been tested
and certified to satisfy SLA objectives.

This feature allows admin to set volume size limit
[per_volume_gigabytes] for a tenant.

Change-Id: Iab584c72a34f9171000b20b2200ac12d21fc40ae
Implements: blueprint cinder-quota-define-per-volume