Updated openstack/openstack Project: openstack/keystone-specs  1dd7960bd4e6bfcc25a8aeac5e64176235528f55
Cleanup and removal of StrictABC requirement

Cleanup the spec (minor wordin changes) and remove the requirement
of the StrictABC module. This simplifies the spec and eliminates what
is likely to be a significant source of bikeshedding and headaches due
to the general rejection of the StrictABC model (fail on class
definition if the method signature do not match).

Change-Id: I00b13d3c551f48074fa74e87c8e86ce24c725195