Updated openstack/openstack Project: openstack-infra/jenkins-job-builder  f2eb6dda32ea2ebfff13a9b4b03bf18ac5f28ef1
Make include_path deterministic

In code the `include_path` and defaults like `.` added to a set as opposed a list.

Resulting in that the order is machine/hashcode specific and
that user cannot make i.e. `.` to have higher precedence than
the fallbacks.

Without doing this you have to use absolute or unique paths which
might not always be possible in your .ini file.

Change-Id: I9721b168a67e799dfc19d7df951b99c0749b52e0