Updated openstack/openstack Project: openstack/governance  8a4a37b9a13ce8f6d1b4bad4fca6b7b24f452cb1
Be specific about geography in release names

The existing relase name criteria were intentionally vague about
geography to allow us to collect the largest amount of cool names
possible for voting.  Somehow, that did not end up happening and
instead a specific determination of region was used to drop about
2/3 of the otherwise-qualifying names from the M name poll after
nominations were closed.

In order to prevent this from happening again, we will need to
very specifically designate a region upon initiating the process.

Also, provide clear indication that the role of the election
official is not to make subjective determinations.

Also, allow for a time between finalizing the list of names and
the start of the election so that if the election official makes
any disputed decisions, there is an opportunity to address them
before the poll goes out.  This also clarifies when the TC should
add any exceptions to the process.

Finally, add an extra column to the election roster to indicate
where the geographic boundaries should be specified.

For more background, see:
http://lists.openstack.org/pipermail/openstack-dev/2015-June/066995.html

Change-Id: Ia9adc66090f2994bb291967f7f2e1b2b4dd14603