Updated openstack/openstack Project: openstack/governance  8f93e1368828f1fc431fd7c811e19c23c3e8a250
Be less proscriptive about new team size/maturity

The new project requirements previously contained prose which was
ambiguous or misleading about when it's okay to apply to the TC for
creation of a project team. New projects with any size team working
on software at any stage of development are welcome to apply.

Change-Id: I8d80681caccee7570303c46fe603426f76254af0