Updated openstack/openstack Project: openstack/nova  38e734e9783f28761e2f6dc4446802af7d7fd6cb
Test relationships of List objects

Originally, List objects were ignored when doing test_relationships().
This is because the relationships of List objects were tested in
child_versions. This worked for List->Obj testing, but did not cover the
Obj->List->Obj relationships. Because of this, the dependency versions
in the List objects became stale.

This change opens up testing on the relationships of the List
objects and corrects the object relationships to include the List
objects.

Change-Id: I62057e2bc1690566d2fa10b41a19fd541d6e42c3
Closes-Bug: #1470153