Updated openstack/openstack Project: openstack-infra/system-config  ef67c6f5efff7d2c468f06571dce51a30dbf1788
Add database backups for openstackid.org

The openstackid.org server uses two databases (one for its own data
and one for the Silverstripe backend to the www site). Keep the
usual 30 days of rotated ID database dumps, but because the SS
content is of significant size just keep a couple.

Change-Id: Id12a02adc57c518eac3facf996f615465a5878ba