Updated openstack/openstack Project: openstack/horizon  9f563c498d6e63e6efd2f5c6f59193905b2a2b5b
Useing 'yesno' filter to display Boolean value

In Project and User detail pages displaying the enabled
boolean field in True and False. Which is not correct or
match with other boolean displays.

This patch using the 'yesno' filter and 'capfirst' filter
to display the value in Yes and No format.

Change-Id: Ib5ffd2b05b5ed59a36571d134fbbe3781b866c84
Closes-Bug: #1469651