Updated openstack/openstack Project: openstack/tempest  f6067a03e7c93e11507c8b1ab89f70ba22c1af5b
Add resource prefix string

When running tempest several resources are created and afterwards
deleted. On systems where the cloud under tests is not destroyed
after the tests, it is useful to be able to associated resources
to specific test jobs and test runs.

The 'rand_name' function in tempest-lib exposes a prefix option,
which tempest does not use at the moment.

Adding a new configuration option to make the prefix configurable.
Creating a local version of the rand_name util, which passes the
prefix from configuration. Eventually all calls of rand_name
shall be redirected from tempest_lib to the local version.
In this patch, only the credentials provider is change to
validate that the mechanism is working. Follow-up patches will
migrate the rest of the code base.

Co-Authored-By: Fei Long Wang (flwang@catalyst.net.nz)

Change-Id: If15993f1ca0fd932e13d73c8ef9875ce13d79c53