Updated openstack/openstack Project: openstack/oslo.config  985ac6257bb1152b93fd31558b19d1c1cd223f72
Add set_default() method to fixture

Add a set_default() method to the fixture to allow it to be passed to
set_default() functions in other oslo libraries that expect to be given
a ConfigObj instance. This lets tests in one project use a fixture to
set defaults for options in code they don't own, and unset those
defaults when the test is cleaned up.

Change-Id: Ifa9e6e2ab4ab4f93a7dc22d786ee6fd6ae7224ae