Updated openstack/openstack Project: openstack-infra/project-config  0e44a35279e4be01dbc4a2cbc9829ea4ffe05e0b
Fix script file name for Inspector dsvm job

Indeed, "functest" is a weird directory for an integration test,
so we're moving it to the devstack plugin directory and also
renaming so that it's purpose is more clear.

Right now we maintain 2 copies of this script, one in
"functest" directory will be deleted from tree once this lands.

Change-Id: I1c9d73be878f14c6a6616b4a93e8552f33d27522