Updated openstack/openstack Project: openstack/heat  6bf1cd9f45e35e063e718fb91e6ca033867823c2
Convergence: Fix for finding best existing res

If we find a resource with template as current template given for
update, then that is the best resource and we should stop continuing.
This happens in case of rollback, when the previous template is used for
updating the stack.

If we encounter a resource with same template as previous template,
we still need to continue to find a possible resource with current
template, to cover the above mentioned case.

Renaming get_existing_db_resource to get_best_existing_db_resource for
better reflection of what it does and moving it out of enclosed method
so that it can be tested.

Change-Id: I5c753e1c0737803ee3de8d36f25c0e620a65f0cf