Updated openstack/openstack Project: openstack/tempest  30859df18b5515979bec3d59b34c472772093e71
Apply a naming rule of GET to secgroup clients

[GET /resources] methods should be "list_<resource name>s"
or "show_<resource name>", so this patch applies the rule
to compute clients which names are "security_group*".

In addition, this patch removes unnecessary str() because
all callers specify string values and we need to pass py3
tests when migrating them into tempest-lib.

Partially implements blueprint consistent-service-method-names

Change-Id: I21f6d3988698e39d69754b58feaacd27d4f76af3