Updated openstack/openstack Project: openstack/nova  e4e16e9077b5e78f809c6c634153fa62a5facb16
Avoid recursion in object relationships test

In the object_relationships test, we try to walk the object tree
to find all dependencies and put them in our mapping. However, we
really don't need to recurse through every relationship to find
all the objects, since they're all in the registry. We really just
need to linearly walk the registry and inspect one level below each
object to find the dependencies. The hint here is that the static
stored mapping is not a tree, it's just a mapping.

Removing the recursion still finds all the same objects and deps,
as evidenced by the fact that this still passes when removing the
recursion step.

Closes-Bug: #1470302
Change-Id: I615d48db5e821e1dc85114277eaf1f6cd5be8622