Updated openstack/openstack Project: openstack/barbican-specs  f31ceb754f6ac958f0a6a3a0e9aee7c2d1e1c49d
Use underscores instead of dashes as word separators in API resource names

Change project-quotas and project-id to project_quotas and project_id.

To be consistent with the OpenStack API WG recommendations: we should
use dashes in the resource names (the request URLs) and use underscores
in the field names (the JSON response).

Relates-to: blueprint quota-support-on-barbican-resources

Change-Id: Icebf327c3c3b52afb0f847d6ab626e6d8eda12aa