Updated openstack/openstack Project: openstack/python-openstackclient  60d1417c2f47006c522c6ed1b973ab3bbb84c69f
Add functional tests for image set

This patch includes functional tests for image set and it includes
a change to use the OSC utils.format_dict method to format the
properties.  This will give a more user friendly format to the image
commands and it gives a more consistent testable format to the
output.  Instead of:

{u'a': u'b', u'c': u'd'}

The user will see:

a=b, c=d

Change-Id: Ib396316586ffc5dbab231064d5b6dc9425507934