Updated openstack/openstack Project: openstack/cinder  d2d48e564d0bf43cd71e86fb6bdf834ff5f278ce
Fix timeout issue in EMC VNX driver unit test

This patch is going to fix the timeout issue in EMC VNX
driver unit test.

Currently, the timeout value of a looping call in EMC unit test
is set to 0.000002s which is too small. The execution time of
a looping call may exceed this timeout value and cause the test
case failure. This fix is changing the timeout to the default
value of VNX driver.

Change-Id: I4aea02403ada8892e45b974be70fb5753721ed39
Closes-Bug: #1467763