Updated openstack/openstack Project: openstack/python-swiftclient  29e5c5656e9da4f2134e754b03fa23410209fdb5
Add connection release test

This patch adds a small test to ensure a connection
is released after all chunks have been consumed.

It's a follow up to commit 8756591b and added to ensure
there will be no regression in the future (this test
fails also with that patch not applied).

Change-Id: I6a6fcd26879eb2070f418c8770a395ff6c30aa51