Updated openstack/openstack Project: openstack/python-swiftclient  f5fa154f6d3ce8833934b6c68257277a2d7174fd
Add passenv to tox.ini to make functests run with tempauth

Since tox version 2.0.0 env vars are not passed to the test
env, which means that the SWIFT_TEST_CONFIG_FILE var is not
passed in to tox -e func env. That means that both times
tox -e func runs it is using keystone auth, and never using
tempauth.

Related-Bug: 1455102

Co-Authored-By: Christian Schwede <cschwede@redhat.com>

Change-Id: I23dcdbcde0bf8adc9429eb2d294a2c778005d136