Updated openstack/openstack Project: openstack/nova  06751d81f330e0f56146f03a395261581bc9fb30
tests: move a test to the proper class in test_resource_tracker

The test test_set_instance_host_and_node very little to do with resizing
an instance in the resource tracker so we move it out of the
ResizeClaimTestCase.

Also test_set_instance_host_and_node does nothing with instance claims,
so it too should live in a more general test suite. Move it too.

Change-Id: Iae7664df6b92b502f2c50799b1101fe7d69dc7b2