Updated openstack/openstack Project: openstack/keystone  0d84009c0cfa07219730973160422bb7cb898f17
Update requirements by hand.

This folds the py3 versions (which infra were not linting until
recently) into the requirements files they should now use.

Note that the -e line for ldap wasn't meant to ever merge, and
I'm not sure it can be merged now, so I've dropped it and we'll
see what happens on test.

Change-Id: Idbd162d70620f683d1bc664193b2a5fe64c159b9