Updated openstack/openstack Project: openstack/horizon  bc0fecd97c22c769b0213f83c1c208932705e66a
Adding identity ng-users panel

This sets up the python side of the users panel.
We need to do this since a panel still needs to inherit from base.html
and navigation is still tightly coupled with Dashboard.

This provides a cleaner way to phase out the old users panel.
This ensures that integration tests does not break until we are ready to
replace the old panel.

To test set DISABLED to False in _250_identity_users_panel.py

Change-Id: Ied6db285ccac0d70a7be4cb1e795b1a2c9058c19
Partially-Implemenents: blueprint angularize-identity-tables