Updated openstack/openstack Project: openstack-infra/project-config  92a72d901eb3c522901cbdc939c45fac09b9ee35
Add openstack/releases repository

This is to be a new, empty, repository used for release requests as
described in https://review.openstack.org/191193. For now there are no
test jobs, but some will be added later. The publish jobs are defined
but not used because we expect to eventually publish some static HTML
out of this repository describing our release history.

Change-Id: Ic821b091ffcba7ebb40309a67d241b1f2fbfba02